Objectify sprite sheet creation
- Added SpriteSheet class:
    * Handle internally the creation/update of the sheet;
    * Have an interal state of the current sheet;
    * Rely on a outside source of sf::Image.

- Current sprite sheet creation:
    * The 6 first frames generated;
    * Saved to file at the 7th frame generated;
    * File is a jpeg with black background;
    * Filename includes a timestamp, the numbers of
    frames and the size of one frame;
    * Stored in the current working directory.

- Corrected as much as possible warnings shown by MSVC;

- Re-added *.lib rule for the .gitignore.

+++ src/TextureManager.cpp
@@ -36,7 +36,8 @@ TextureManager::TextureManager(const char *filename)
 		_safeScale = (displayRatio > static_cast<float>(size.x) / size.y ?
 			       static_cast<float>(desktopMode.height) / size.y
 			       : static_cast<float>(desktopMode.width) / size.x);
-		_safeSize = sf::Vector2u(size.x * _safeScale, size.y * _safeScale);
+		_safeSize = sf::Vector2u(static_cast<unsigned int>(size.x * _safeScale),
+					 static_cast<unsigned int>(size.y * _safeScale));
 	} else {
 		_safeScale = 1.0f;
 		_safeSize = _texture.getSize();
@@ -59,8 +60,10 @@ sf::Sprite TextureManager::getPartialTextureAsSprite(sf::IntRect dimensions) con
 {
 	sf::Sprite sprite(_texture);
 	// Deduce the original positionx from the scaled down region.
-	sf::IntRect scaled(dimensions.left / _safeScale, dimensions.top / _safeScale,
-			   dimensions.width / _safeScale, dimensions.height / _safeScale);
+	sf::IntRect scaled(static_cast<int>(dimensions.left   / _safeScale),
+			   static_cast<int>(dimensions.top    / _safeScale),
+			   static_cast<int>(dimensions.width  / _safeScale),
+			   static_cast<int>(dimensions.height / _safeScale));
 
 	sprite.setTextureRect(scaled);
 	sprite.setScale(_safeScale, _safeScale);
